use leaflet instead of d3 for mapping
This is a sketch of how we might use the Leaflet library instead of d3
for displaying the map. The benefits are that the clustering and the
zooming are already implemented in Leaflet, so we avoid having to
create a robust interactive map from d3's primitives.

d3 is still used for data manipulation and will be useful for other
parts of the data display.

To avoid repeated calls to the geocoding API, for this sketch I've
included the data from one round of calls to the Nominatim service.